[SD][web] Add Img2Img UI
