+++ apps/stable_diffusion/web/ui/__init__.py
@@ -1,0 +1,2 @@
+from apps.stable_diffusion.web.ui.txt2img_ui import txt2img_web
+from apps.stable_diffusion.web.ui.img2img_ui import img2img_web
