+++ apps/stable_diffusion/web/ui/css/sd_dark_theme.css
