+++ apps/stable_diffusion/web/ui/logos/nod-logo.png
