+++ apps/stable_diffusion/web/ui/logos/sd-demo-logo.png
